Clean Database & Run Seeder

+++ app/Database/Migrations/2021-06-12-115225_CreatePostTable.php
@@ -44,7 +44,7 @@ class CreatePostTable extends Migration
 				'constraint' => '11',
 			],
 			'price'       => [
-				'type'       => 'BIGINT',
+				'type'       => 'INT',
 				'constraint' => '11',
 				'null' => true,
 				'default'    => '0',
